Removed hello controller and renamed controller package
